Переработал паттерн страницы /administration
