+++ page_objects/admin_page/admin_login_page.py
@@ -1,12 +1,11 @@
 from selenium.webdriver.common.by import By
-from page_objects.base_page import BasePage
+from page_objects.admin_page.navigation_admin_menu_page import NavigationAdminMenuPage
 
 
-class AdminPage(BasePage):
+class AdminLoginPage(NavigationAdminMenuPage):
     USERNAME_INPUT = By.ID, 'input-username'
     PASSWORD_INPUT = By.ID, 'input-password'
     LOGIN_BUTTON = By.CSS_SELECTOR, '.text-end button'
-    LOGOUT_BUTTON = By.CSS_SELECTOR, '.fa-sign-out'
     ADMIN_URL_PAGE = '/administration'
 
     def find_password_input(self):
@@ -23,9 +22,3 @@ class AdminPage(BasePage):
 
     def click_login_button(self):
         self.click(self.LOGIN_BUTTON)
-
-    def find_total_order(self, page: str):
-        self.click((By.XPATH, self._text_xpath(page)))
-
-    def click_logout_button(self):
-        self.click(self.LOGOUT_BUTTON)
